moving styles to checkbox horizontal field and renaming files, used field
